refactor: FE config, res path, env use, file name
