+++ backend/src/utils/secrets.utils.js
