+++ frontend/src/app/shared/interfaces/auth.interface.ts
@@ -2,5 +2,5 @@ export declare interface AuthResponse {
     user: string,
     user_id: string,
     expiresIn: string,
-    token: string
+    token: any
 }
